Split up test_views
This will make it easier to run tests for a single view
